Lift 2.12 Scala Native restriction
Not sure if a particular Scala 2.12 version lifted that issue, but it
sure works with the latest Scala Native and 2.12 versions.
