+++ modules/integration/src/test/scala/scala/cli/integration/RunTestDefinitions.scala
@@ -132,9 +132,6 @@ abstract class RunTestDefinitions(val scalaVersionOpt: Option[String])
 
   def platformNl: String = if (Properties.isWin) "\\r\\n" else "\\n"
 
-  def canRunScWithNative: Boolean =
-    !(actualScalaVersion.startsWith("2.12") || actualScalaVersion.startsWith("3.0"))
-
   def simpleNativeTests(): Unit = {
     val fileName = "simple.sc"
     val message  = "Hello"
@@ -159,38 +156,9 @@ abstract class RunTestDefinitions(val scalaVersionOpt: Option[String])
     }
   }
 
-  if (canRunScWithNative)
-    test("simple script native") {
-      simpleNativeTests()
-    }
-  else
-    test("Descriptive error message for unsupported native/script configurations") {
-      val inputs = TestInputs(
-        Seq(
-          os.rel / "a.sc" -> "println(1)"
-        )
-      )
-      val nativeVersion = "0.4.2"
-      inputs.fromRoot { root =>
-        val output = os.proc(
-          TestUtil.cli,
-          extraOptions,
-          "--native",
-          "a.sc",
-          "--native-version",
-          nativeVersion
-        ).call(
-          cwd = root,
-          check = false,
-          stderr = os.Pipe
-        ).err.text().trim
-        expect(
-          output.contains(
-            s"Used Scala Native version $nativeVersion is incompatible with Scala $actualScalaVersion."
-          )
-        )
-      }
-    }
+  test("simple script native") {
+    simpleNativeTests()
+  }
 
   if (actualScalaVersion.startsWith("3.1"))
     test("Scala 3 in Scala Native") {
@@ -324,10 +292,9 @@ abstract class RunTestDefinitions(val scalaVersionOpt: Option[String])
     }
   }
 
-  if (canRunScWithNative)
-    test("Multiple scripts native") {
-      multipleScriptsNative()
-    }
+  test("Multiple scripts native") {
+    multipleScriptsNative()
+  }
 
   test("Directory") {
     val message = "Hello"
