+++ modules/integration/src/test/scala/scala/cli/integration/RunTests212.scala
@@ -1,7 +1,38 @@
 package scala.cli.integration
 
+import com.eed3si9n.expecty.Expecty.expect
+
 // format: off
 class RunTests212 extends RunTestDefinitions(
   scalaVersionOpt = Some(Constants.scala212)
-)
-// format: on
+) {
+  // format: on
+
+  test("Descriptive error message for unsupported native/script configurations") {
+    val inputs = TestInputs(
+      Seq(
+        os.rel / "a.sc" -> "println(1)"
+      )
+    )
+    val nativeVersion = "0.4.2"
+    inputs.fromRoot { root =>
+      val output = os.proc(
+        TestUtil.cli,
+        extraOptions,
+        "--native",
+        "a.sc",
+        "--native-version",
+        nativeVersion
+      ).call(
+        cwd = root,
+        check = false,
+        stderr = os.Pipe
+      ).err.text().trim
+      expect(
+        output.contains(
+          s"Used Scala Native version $nativeVersion is incompatible with Scala $actualScalaVersion."
+        )
+      )
+    }
+  }
+}
